Update index dot md
Update Read File md, and others
