+++ docs/_includes/brackets-ide.md
@@ -1,6 +1,6 @@
 ## Brackets IDE
 
-The [Brackets IDE](http://brackets.io/){:title="Click to download the Brackets IDE"}{:target="_blank"} allows the Javascript developer to display instant changes to both your Javascript program (.js) and your View page (.html) through the Chrome browser window.
+The [Brackets IDE](https://brackets.io/){:title="Click to download the Brackets IDE"}{:target="_blank"} allows the Javascript developer to display instant changes to both your Javascript program (.js) and your View page (.html) through the Chrome browser window.
 
 **Note**. The Javascript console may also be exposed through Chrome's developer tools while viewing the Chrome browser window.
