+++ docs/_layouts/default.htm
@@ -8,7 +8,7 @@
       
       <h1 class="project-name">{{ site.title | default: site.github.repository_name }}</h1>
 
-      <h2 class="project-tagline">{{ site.description | default: site.github.project_tagline }}</h2>
+      <h2 class="project-tagline">{{ site.tagline | default: site.github.project_tagline }}</h2>
 
       {% if site.github.is_project_page %}
         <a href="{{ site.github.repository_url }}" class="btn">View on GitHub</a>
