+++ docs/pages/App-One.md
@@ -1,0 +1,34 @@
+---
+title: App One
+layout: default
+navigation_weight: 2
+---
+# App One
+
+{{ site.tagline }}
+
+{% include toc-flammarion.md %}
+
+## Subtitle Begin Template Dtd 12-31-17
+
+Place the introducing line of text ie.) the 'tagline' here ...
+
+## Last Subtitle
+
+Place the introducing line of text ie.) the 'tagline' here ...
+
+```liquid
+{% raw %}
+Enjoy the successful output!
+{% endraw %}
+```
+
+{% include brackets-ide.md %}
+
+{% include sources-and-uses.md %}
+
+### External Sources
+
+- The [Node.js™ Course](https://www.w3schools.com/nodejs/){:title="Click to Visit the Node.js™ Course at W 3 Schools dot com"}{:target="_blank"} at W 3 Schools dot com. Published by © 2017 [W3schools.com](https://www.w3schools.com/){:title="Click to Visit the Home page of W 3 Schools dot com"}{:target="_blank"}.
+
+- The [Project Source Links](https://mminail.github.io/Node/Source-Node-Links.htm){:title="Click to Visit the Source Links page of the Node Lessons Project at GitHub pages"}{:target="_blank"} page of the Node Lessons Project. Published by © 2000 - 2018 [Mminail.github.io](https://mminail.github.io/){:title="Click to Visit the Concept Library of the Medical Marijuana Initiative of North America - International Limited, an Arizona Benefit Corporation"}{:target="_blank"}.
