+++ docs/pages/Express-Hello.md
@@ -1,0 +1,34 @@
+---
+title: Express Hello
+layout: default
+navigation_weight: 3
+---
+# Express Hello
+
+{{ site.tagline }}
+
+{% include toc-flammarion.md %}
+
+## Subtitle Begin Template Dtd 12-31-17
+
+Place the introducing line of text ie.) the 'tagline' here ...
+
+## Last Subtitle
+
+Place the introducing line of text ie.) the 'tagline' here ...
+
+```liquid
+{% raw %}
+Enjoy the successful output!
+{% endraw %}
+```
+
+{% include brackets-ide.md %}
+
+{% include sources-and-uses.md %}
+
+### External Sources
+
+- The [Project Source Links](https://mminail.github.io/Node/Source-Node-Links.htm){:title="Click to Visit the Source Links page of the Node Lessons Project at GitHub pages"}{:target="_blank"} page of the Node Lessons Project. Published by © 2017 [Mminail.github.io](https://mminail.github.io/){:title="Click to Visit the Concept Library of the Medical Marijuana Initiative of North America - International Limited, an Arizona Benefit Corporation"}{:target="_blank"}.
+
+- The [Node Express Hello Program](https://www.frontendmasters.com/courses/web-development/express){:title="Click to Visit the 'Node Express Hello Program' of the Express Web-development Course at Front End Masters dot com"}{:target="_blank"} of the Express Web-development Course at Front End Masters dot com. Published by © 2017 [Frontendmasters.com](https://www.frontendmasters.com/){:title="Click to Visit the Home page of Front End Masters dot com"}{:target="_blank"}.
