+++ docs/pages/Node-Package-Mgr.md
@@ -1,7 +1,7 @@
 ---
 title: Node Package Mgr
 layout: default
-navigation_weight: 9
+navigation_weight: 8
 ---
 # Node Package Mgr
 
@@ -29,10 +29,22 @@ npm i -g npm@latest
 {% endraw %}
 ```
 
+## Last Subtitle
+
+Place the introducing line of text ie.) the 'tagline' here ...
+
+```liquid
+{% raw %}
+Enjoy the successful output!
+{% endraw %}
+```
+
+{% include brackets-ide.md %}
+
 {% include sources-and-uses.md %}
 
 ### External Sources
 
-- The [Project Source Links](https://mminail.github.io/Node/Source-Node-Links.htm){:title="Click to Visit the Source Links page of the Node Lessons Project at GitHub pages"}{:target="_blank"} page of the Node Lessons Project. Published by © 2017 [Mminail.github.io](https://mminail.github.io/){:title="Click to Visit the Concept Library of the Medical Marijuana Initiative of North America - International Limited, an Arizona Benefit Corporation"}{:target="_blank"}.
+- The [Project Source Links](https://mminail.github.io/Node/Source-Node-Links.htm){:title="Click to Visit the Source Links page of the Node Lessons Project at GitHub pages"}{:target="_blank"} page of the Node Lessons Project. Published by © 2000 - 2018 [Mminail.github.io](https://mminail.github.io/){:title="Click to Visit the Concept Library of the Medical Marijuana Initiative of North America - International Limited, an Arizona Benefit Corporation"}{:target="_blank"}.
 
 - [The Complete Guide to Angular by Nate Murray, Felipe Coury, Ari Lerner, and Carlos Taborda](https://www.ng-book.com/2){:title="Click to Visit The Complete Guide to Angular aka ng-book 2"}{:target="_blank"}. Published by © 2017 [Fullstack.io](https://www.fullstack.io){:title="Click to Visit The Complete Guide to Angular"}{:target="_blank"}.
