+++ src/Appone/js/app-one.js
@@ -1,0 +1,15 @@
+// Appone
+/// @title App One = Appone
+// - A simple program written in Javascript that takes advantage of the Node.js™ Server program.
+// Copyright (c) 2000 - 2018 Robert Weber FBO Node Lessons Project
+// C=> https://rwebaz.github.io/Node-Lessons-Project/pages/App-One.html
+
+"use strict";
+// As of ES7 the Global use of the strict string pragma is recommended
+
+hljs.initHighlightingOnLoad();
+/* $(document).ready(function() {
+  $('pre code').each(function(i, block) {
+    hljs.highlightBlock(block);
+  });
+}); */
