+++ src/Nodedate/js/node-date.js
@@ -1,0 +1,12 @@
+// Nodedate
+/// @title Node Date = Nodedate
+// - A simple program written in Javascript that takes advantage of the Node.js™ Server program.
+// Copyright (c) 2000 - 2018 Robert Weber FBO Node Lessons Project
+// C=> https://rwebaz.github.io/Node-Lessons-Project/pages/Node-Date.html
+
+"use strict";
+// As of ES7 the Global use of the strict string pragma is recommended
+
+exports.myDateTime = function () {
+    return Date();
+};
