+++ src/Nodehello/js/node-hello.js
@@ -1,4 +1,11 @@
-// Node Hello C=> https://www.w3schools.com/nodejs/nodejs_get_started.asp
+// Nodehello
+/// @title Node Hello = Nodehello
+// - A simple program written in Javascript that takes advantage of the Node.js™ Server program.
+// Copyright (c) 2000 - 2018 Robert Weber FBO Node Lessons Project
+// C=> https://rwebaz.github.io/Node-Lessons-Project/pages/Node-Hello.html
+
+"use strict";
+// As of ES7 the Global use of the strict string pragma is recommended
 
 // Load the required modules
 var http = require('http');
