+++ src/Readfile/index.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<!-- Set the base language to English -->
+<html lang="en-US">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width,initial-scale=1">
+    <title>Read File</title>
+  </head>
+
+  <body>
+    <hgroup class='text-left'>
+      <h4 id="title">${page.title}</h4>
+    </hgroup>
+      <hr class='green-groove' />
+
+    <div id="content">...</div>
+
+<!-- Place all scripts above the closing body tag -->
+<script src="../js/read-file.js"></script> -->
+
+  </body>
+</html>
