--- src/app-one/js/node-date.js
@@ -1,5 +1,0 @@
-// Node Date Module C=> https://www.w3schools.com/nodejs/nodejs_modules.asp
-
-exports.myDateTime = function () {
-    return Date();
-};
