--- src/app-one/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "app-one",
-  "version": "0.0.1",
-  "description": " A crypto currency interface",
-  "main": "app.js",
-  "directories": {
-    "doc": "docs"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/rwebaz/Node-Lessons-Project.git"
-  },
-  "author": "Robert Weber",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/rwebaz/Node-Lessons-Project/issues"
-  },
-  "homepage": "https://github.com/rwebaz/Node-Lessons-Project#readme",
-  "dependencies": {
-    "express": "4.16.2"
-  },
-  "devDependencies": {},
-  "keywords": [
-    "cryptocurrency"
-  ]
-}
